paicoin/requirements.txt
APScheduler==3.6.3
boto3==1.13.19
botocore==1.16.19
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
docutils==0.15.2
Flask==1.1.2
idna==2.9
inotify==0.2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
nose==1.3.7
python-bitcoinrpc==1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
tzlocal==2.1
urllib3==1.25.9
Werkzeug==1.0.1
